requirements.txt
# Essential AI/ML and Development libraries
numpy
pandas
scikit-learn
python-dotenv
requests
pytest
loguru
pypdf2
pdfplumber
python-docx
openai
langchain
